Splited treatment of relational data, improved roles to support read and write permissions
