+++ test/configs/models/families.js
@@ -82,7 +82,8 @@ export default {
     }
   ],
   "visible": true,
-  "roles": [
-    "all"
-  ]
+  "roles": {
+    "read": ["all"],
+    "write": ["all"]
+  }
 }
