+++ test/configs/models/images.js
@@ -77,7 +77,8 @@ export default {
     }
   ],
   "visible": true,
-  "roles": [
-    "all"
-  ]
+  "roles": {
+    "read": ["all"],
+    "write": ["all"]
+  }
 }
