+++ test/configs/models/roles.js
@@ -62,7 +62,8 @@ export default {
     }
   ],
   "visible": true,
-  "roles": [
-    "all"
-  ]
+  "roles": {
+    "read": ["all"],
+    "write": ["all"]
+  }
 }
